refactor: 변경 BaseEntity 패키지 경로
Fixes: #2
